fix: add validation for missing IDs in various API handlers and improve type definitions and user caching based on the session_id

+++ server/api/articles/[id]/index.delete.ts
@@ -1,6 +1,7 @@
 export default defineMultiFactorVerificationEventHandler(async (event, { server }) => {
 
     const id = getRouterParam(event, "id")
+    if (!id) return useReturnResponse(event, badRequestError);
 
     const { data: attachments, error: fetchError } = await server.from('attachments').select('name').eq('article_id', id)
     const { error } = await server.from('artikelen').delete().eq('id', id);
